+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Toggle button component
 * Multiple files selection with image picker
 * Date input component
+* Switcher component
 
 ### Changed